Add iterator for objects, change null to error on strict
This commit mainly adds the ability to iterate over objects with the
iterator operator as described in the jq manual, in the same fashion as
arrays can be iterated over.

Additionally the handling for iterating null has changed to be more
jq-specific, treating it like other unsupported types and specifically
identifying null as "null" and not "object".