Improvement for iterators
This ensures that micro-jq conforms to jq's "spec" of constructing
multiple objects when a field in an object constructor produces
"multiple results".

The text in the manual is slightly ambiguous to the actual behavior
here: "multiple results" generally refers to when an iterator
operation is called in the sub-expression for a particular object field,
versus *any* result, such as an expression that simply returns an array.
Hence, while "{ foo: .[] }" being  called on "[1, 2]" would return two
objects with a key of "foo" and a value of 1 and 2 respectively,
"{ foo: . }" itself without the iterator would return a single object
with the "foo" field pointing to the array itself. This behavior carries
over to any downstream expression as well, such as in the event of a
pipe.

jq accomplishes this behavior via the use of backtracking, generators,
and most specifically "fork points" (see
https://github.com/stedolan/jq/wiki/Internals:-backtracking). This
approach is significantly more complex than micro-jq's at this time,
and adopting something akin to it would probably require significant
rewriting of the interpreter and the grammar internals. Ultimately,
however, in regards to where it currently matters to micro-jq, it seems
to boil down to:

* Ensure we can determine when an explode operation occurred. This
commit accomplishes this by adding an optional callback to
evaluateOpCodes. The purpose of this callback can be extended in the
future, but for now, we are using it to signal when an iterator
operation was encountered. The callback also tracks the length of the
iterated Array so that we can handle it properly in the event of zero or
single-length arrays, which may end up getting auto-promoted by
evaluateOpCodes in a way that can affect iteration.

* Use the current behavior for object construction to instead build a
matrix of indexed value operations: a key now creates one or more
iterations of an object now with different individual values. Objects
are then created by applying a simple recursive function to create the
correct amount of objects based on the product of the number of values
each key holds.

* Since pipes are also affected by this, some updates to the parser have
been necessary to explicitly distinguish pipes as an operation. On the
interpreter side we now apply the sub-expression on the LHS of the pipe,
detect if an iterator operation was encountered, and if so, process the
RHS on each element in the result individually.

Since this increases the LOC of the interpreter significantly, this
commit also breaks up the eval operation into several sub-functions,
generally one for each operation.