Don't hand off exploder callback during array creation (#8)
This fixes array creation so that the exploder callback chain is broken
during the operation.

The reason for this is that while multiple results are generated during
object creation, all results from an expression in array creation are
concatenated, with the result being a single array instance. Hence, any
parent operations should just see a single generated value and should
not iterate over its results.